projects/webinar_0313/inventory-agent-app/requirements.txt
# AI エージェント依存関係
mlflow[databricks]>=2.20.0
openai-agents>=0.0.5
databricks-sdk>=0.40.0
pandas>=2.0.0
numpy>=1.24.0
